cookbot-backend-recipe-ingestion/requirements.txt
# web framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# config
pydantic==2.9.2
pydantic-settings==2.5.2

# db
asyncpg==0.29.0
redis[hiredis]==5.0.8

# ingestion
httpx==0.27.2
recipe-scrapers==15.0.0
extruct==0.17.0
beautifulsoup4==4.12.3

# utils
python-slugify==8.0.4
python-json-logger==2.0.7

# dev
pytest==8.3.3
pytest-asyncio==0.24.0
